x-ray-report-generator/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
together
python-dotenv
fpdf